+++ src/cp1610.ts
@@ -1,86 +1,8 @@
+import {Bus, BusDevice} from "./Bus";
+import {Mnemonic, InstructionConfig} from "./InstructionConfig";
 import {UnreachableCaseError} from "./UnreachableCaseError";
-import instructions from "./instructions";
-
-let totalLogs = 0;
-const trace = (..._: any[]) => {
-  if (totalLogs < 359) {
-    console.error(..._);
-    totalLogs += 1;
-  }
-};
-
-export type InstructionConfig = Readonly<{
-  instruction: string;
-  mnemonic: Mnemonic;
-  cycles: ReadonlyArray<number>;
-  interruptable: boolean;
-  in_s?: boolean;
-  in_z?: boolean;
-  in_o?: boolean;
-  in_c?: boolean;
-  in_d?: boolean;
-  out_s?: boolean;
-  out_z?: boolean;
-  out_o?: boolean;
-  out_c?: boolean;
-  out_i?: boolean;
-  out_d?: boolean;
-}>;
-
-export type Mnemonic =
-  | "HLT"
-  | "SDBD"
-  | "EIS"
-  | "DIS"
-  | "J"
-  | "TCI"
-  | "CLRC"
-  | "SETC"
-  | "INCR"
-  | "DECR"
-  | "COMR"
-  | "NEGR"
-  | "ADCR"
-  | "GSWD"
-  | "NOP"
-  | "SIN"
-  | "RSWD"
-  | "SWAP"
-  | "SLL"
-  | "RLC"
-  | "SLLC"
-  | "SLR"
-  | "SAR"
-  | "RRC"
-  | "SARC"
-  | "MOVR"
-  | "ADDR"
-  | "SUBR"
-  | "CMPR"
-  | "ANDR"
-  | "XORR"
-  | "B"
-  | "MVO"
-  | "MVO@"
-  | "MVOI"
-  | "MVI"
-  | "MVI@"
-  | "MVII"
-  | "ADD"
-  | "ADD@"
-  | "ADDI"
-  | "SUB"
-  | "SUB@"
-  | "SUBI"
-  | "CMP"
-  | "CMP@"
-  | "CMPI"
-  | "AND"
-  | "AND@"
-  | "ANDI"
-  | "XOR"
-  | "XOR@"
-  | "XORI";
+import {decodeOpcode} from "./decodeOpcode";
+import {trace} from "./trace";
 
 /**
  * Internal state of the CPU. This is essentially half of a state machine; the
@@ -241,179 +163,7 @@ const INSTRUCTION_STATES: Record<Mnemonic, CpuState> = {
   XORI: "INDIRECT_READ:BAR",
 };
 
-export interface BusDevice {
-  clock(): void;
-  debug_read(addr: number): null | number;
-}
-
-export type BusFlags =
-  | 0b000
-  | 0b001
-  | 0b010
-  | 0b011
-  | 0b100
-  | 0b101
-  | 0b110
-  | 0b111;
-
-const BUS_FLAG_STRINGS = [
-  "NACT",
-  "ADAR",
-  "IAB",
-  "DTB",
-  "BAR",
-  "DW",
-  "DWS",
-  "INTAK",
-] as const;
 type RegisterIndex = 0 | 1 | 2 | 3 | 4 | 5 | 6 | 7;
-
-export class Bus {
-  /**
-   * NACT - No ACTion
-   *
-   * - BDIR: 0
-   * - BC2: 0
-   * - BC1: 0
-   *
-   * During this stage, no device is active on the bus. DB0 through DB15 are
-   * allowed to float, with their previous driven value fading away during this
-   * phase.
-   */
-  static ___ = 0b000 as const;
-  /**
-   * ADAR - Address Data to Address Register
-   *
-   * - BDIR: 0
-   * - BC2: 0
-   * - BC1: 1
-   *
-   * This bus phase is issued by the CPU during a Direct Addressing Mode
-   * instruction. Prior to this phase, an address will have been latched in a
-   * device by a prior BAR or ADAR bus phase. Then, during this phase, the
-   * currently selected device responds with its data on the bus, and at the end
-   * of this phase, all devices should latch this address as the address for the
-   * next memory access (DTB, DW, or DWS phases). The CPU asserts nothing during
-   * this phase -- rather, it expects the currently addressed device to inform
-   * the rest of the machine of the address for the next access.
-   */
-  static ADAR = 0b001 as const;
-  /**
-   * IAB - Interrupt Address to Bus
-   *
-   * - BDIR: 0
-   * - BC2: 1
-   * - BC1: 0
-   *
-   * This bus phase is entered during interrupt processing, after the current
-   * program counter has been written to the stack. It's also entered into on
-   * the first cycle after coming out of RESET. During this phase, an external
-   * device should assert the address of the Interrupt or RESET vector as
-   * appropriate. The CPU then moves this address into the program counter and
-   * resumes execution.
-   */
-  static IAB = 0b010 as const;
-  /**
-   * DTB - Data To Bus
-   *
-   * - BDIR: 0
-   * - BC2: 1
-   * - BC1: 1
-   *
-   * This phase is entered during a read cycle. During this phase, the currently
-   * addressed device should assert its data on the bus. The CPU then reads this
-   * data.
-   */
-  static DTB = 0b011 as const;
-  /**
-   * BAR - Bus to Address Register
-   *
-   * - BDIR: 1
-   * - BC2: 0
-   * - BC1: 0
-   *
-   * During this phase, the CPU asserts the address for the current memory
-   * access. All devices on the bus are expected to latch this address and
-   * perform address decoding at this time.
-   */
-  static BAR = 0b100 as const;
-  /**
-   * DW - Data Write
-   *
-   * - BDIR: 1
-   * - BC2: 0
-   * - BC1: 1
-   *
-   * The DW and DWS bus phases initiate a write cycle. They always occur
-   * tofETCHher on adjacent cycles, with data remaining stable on the bus across
-   * the transition from DW to DWS. During these phases, the data being written
-   * is available for external memories to latch. The CP-1600 allows two full
-   * CPU cycles for external RAM to latch the data.
-   */
-  static DW = 0b101 as const;
-  /**
-   * DWS - Data Write Strobe
-   *
-   * - BDIR: 1
-   * - BC2: 1
-   * - BC1: 0
-   *
-   * Behaves the same as DW
-   */
-  static DWS = 0b110 as const;
-  /**
-   * INTAK - INTerrupt AcKnowledge
-   *
-   * - BDIR: 1
-   * - BC2: 1
-   * - BC1: 1
-   *
-   * The CPU enters this bus phase on the first cycle of interrupt processing.
-   * During the phase, the CPU places the current stack pointer value on the bus
-   * as it prepares to "push" the current program counter on the stack. Devices
-   * are expected to treat INTAK similarly to a BAR bus phase. Indeed, on the
-   * Intellivision Master Component, only the 16-bit System RAM sees the INTAK
-   * bus phase. It uses this bus phase to trigger a special bus-copy mode as
-   * well as for latching the current address. For all other devices in the
-   * system, INTAK is remapped to BAR by some discrete logic, and so is
-   * processed as a normal addressing cycle elsewhere.
-   */
-  static INTAK = 0b111 as const;
-  ticks: number = 0;
-
-  toString(): string {
-    return `Bus(${BUS_FLAG_STRINGS[this.flags]}:$${this._data
-      .toString(16)
-      .padStart(4, "0")})`;
-  }
-
-  _data: number = 0x0000;
-
-  /**
-   * Current 16-bit data on the bus
-   */
-  get data(): number {
-    return this._data & 0xffff;
-  }
-  set data(_data: number) {
-    // trace(`bus.data = $${_data.toString(16)}`);
-    this._data = _data & 0xffff;
-  }
-
-  //
-  // BUS CONTROL FLAGS
-  //
-  flags: BusFlags = Bus.___;
-
-  clock() {
-    this.ticks = (this.ticks + 1) % 4;
-    // Bus reads 0xFFFF when there's nothing asserting data; this is mainly to
-    // match jzIntv behavior:
-    if (this.ticks === 3 && this.flags === Bus.___) {
-      this._data = 0xFFFF;
-    }
-  }
-}
 
 /**
  * Emulation of the General Instruments CP1610 microprocessor, as used in the
@@ -1622,20 +1372,6 @@ export class CP1610 implements BusDevice {
   }
 }
 
-const opcodeLookup: InstructionConfig[] = [];
-for (const [rangeKey, instructionConfig] of Object.entries(instructions)) {
-  const [start, end = start] = rangeKey
-    .split("-")
-    .map((str) => parseInt(str, 16)) as [number, number | undefined];
-  for (let i = start; i <= end; ++i) {
-    opcodeLookup[i] = instructionConfig;
-  }
-}
-
-export const decodeOpcode = (opcode: number): InstructionConfig | null => {
-  return opcodeLookup[opcode] ?? null;
-};
-
 const decodeRegisterIndex = (opcode: number): RegisterIndex => {
   return (opcode & 0b0000_0000_0000_0111) as any;
 };
@@ -1643,168 +1379,3 @@ const decodeRegisterIndex = (opcode: number): RegisterIndex => {
 export const forTestSuite = {
   decodeOpcode,
 };
-
-export class RAM implements BusDevice {
-  data: Uint16Array;
-  start: number;
-  _addr: number | null = null;
-  bus: Bus;
-  ticks: number = 0;
-  name: string = "RAM";
-
-  constructor(bus: Bus, start: number, end: number) {
-    this.bus = bus;
-    this.start = start;
-    this.data = new Uint16Array(end - start);
-  }
-
-  debug_read(addrIn: number): number | null {
-    const addr = addrIn - this.start;
-    const data = this.data[addr];
-    if (data == null) return null;
-    return data;
-  }
-
-  _readAndDecodeAddr() {
-    const addr = this.bus.data - this.start;
-
-    if (addr >= 0 && addr < this.data.length) {
-      this._addr = addr;
-      trace(
-        `[${this.name}] this._addr = $${this._addr
-          .toString(16)
-          .padStart(4, "0")} (this.bus.data($${this.bus.data
-          .toString(16)
-          .padStart(4, "0")}) - this.start($${this.start
-          .toString(16)
-          .padStart(4, "0")}))`,
-      );
-    } else {
-      this._addr = null;
-    }
-  }
-
-  _assertDataAtAddrToBus() {
-    if (this._addr == null) return;
-
-    const data = this.data[this._addr];
-    if (data == null) return;
-    trace(
-      `[${this.name}] this.bus.data = $${data
-        .toString(16)
-        .padStart(4, "0")} (this._addr($${this._addr
-        .toString(16)
-        .padStart(4, "0")}))`,
-    );
-    this.bus.data = data;
-    this._addr = null;
-  }
-
-  _readDataOnBusToAddr() {
-    if (this._addr == null) return;
-
-    this.data[this._addr] = this.bus.data;
-  }
-
-  clock(): void {
-    // trace(this.name);
-    this.ticks = (this.ticks + 1) % 4;
-
-    switch (this.bus.flags) {
-      case Bus.BAR: {
-        // During this phase, the CPU asserts the address for the current memory
-        // access. All devices on the bus are expected to latch this address and
-        // perform address decoding at this time.
-
-        // Latch the decoded address from the bus if it falls within our address
-        // range:
-        if (this.ticks === 3) return this._readAndDecodeAddr();
-        return;
-      }
-      case Bus.DTB: {
-        // This phase is entered during a read cycle. During this phase, the
-        // currently addressed device should assert its data on the bus. The CPU
-        // then reads this data.
-        if (this.ticks === 1) return this._assertDataAtAddrToBus();
-        return;
-      }
-      case Bus.ADAR: {
-        // This bus phase is issued by the CPU during a Direct Addressing Mode
-        // instruction. Prior to this phase, an address will have been latched
-        // in a device by a prior BAR or ADAR bus phase.
-        //
-        // Then, during this phase, the currently selected device responds with
-        // its data on the bus, and at the end of this phase, all devices should
-        // latch this address as the address for the next memory access (DTB,
-        // DW, or DWS phases).
-        //
-        // The CPU asserts nothing during this phase -- rather, it expects the
-        // currently addressed device to inform the rest of the machine of the
-        // address for the next access.
-        if (this.ticks === 1) return this._assertDataAtAddrToBus();
-        if (this.ticks === 3) return this._readAndDecodeAddr();
-        return;
-      }
-      case Bus.DW: {
-        return;
-      }
-      case Bus.DWS: {
-        // The DW and DWS bus phases initiate a write cycle. They always occur
-        // together on adjacent cycles, with data remaining stable on the bus
-        // across the transition from DW to DWS. During these phases, the data
-        // being written is available for external memories to latch. The
-        // CP-1600 allows two full CPU cycles for external RAM to latch the
-        // data.
-        if (this.ticks === 3) return this._readDataOnBusToAddr();
-        return;
-      }
-      case Bus.IAB: {
-        // This bus phase is entered during interrupt processing, after the
-        // current program counter has been written to the stack. It's also
-        // entered into on the first cycle after coming out of RESET. During
-        // this phase, an external device should assert the address of the
-        // Interrupt or RESET vector as appropriate. The CPU then moves this
-        // address into the program counter and resumes execution.
-        if (this.ticks === 1) return this._assertDataAtAddrToBus();
-        return;
-      }
-      case Bus.INTAK: {
-        // The CPU enters this bus phase on the first cycle of interrupt
-        // processing. During the phase, the CPU places the current stack
-        // pointer value on the bus as it prepares to "push" the current program
-        // counter on the stack. Devices are expected to treat INTAK similarly
-        // to a BAR bus phase. Indeed, on the Intellivision Master Component,
-        // only the 16-bit System RAM sees the INTAK bus phase. It uses this bus
-        // phase to trigger a special bus-copy mode as well as for latching the
-        // current address. For all other devices in the system, INTAK is
-        // remapped to BAR by some discrete logic, and so is processed as a
-        // normal addressing cycle elsewhere.
-        return;
-      }
-      case Bus.___: {
-        // During this stage, no device is active on the bus. DB0 through DB15
-        // are allowed to float, with their previous driven value fading away
-        // during this phase.
-        return;
-      }
-      default: {
-        throw new UnreachableCaseError(this.bus.flags);
-      }
-    }
-  }
-}
-
-export class ROM extends RAM {
-  name: string = "ROM";
-
-  constructor(bus: Bus, start: number, data: Uint16Array) {
-    super(bus, start, start + data.length);
-    // Marginally wasteful since we throw out the original `data` - yeah yeah,
-    // OOP is bad yadda yadda
-    this.data = data;
-  }
-
-  _readDataOnBusToAddr() {
-    // Do nothing. Hence, read-only.
-  }
-}
