+++ src/trace.ts
@@ -1,0 +1,7 @@
+let totalLogs = 0;
+export const trace = (..._: any[]) => {
+  if (totalLogs < 359) {
+    console.error(..._);
+    totalLogs += 1;
+  }
+};
